small local changes for reproduction of results turned into larger work, fixing dependencies for sktime 0.3.0 (previously 0.2.0)
